+++ sanity/schemaTypes/index.ts
@@ -4,5 +4,14 @@ import {categoryType} from './categoryType'
 import body from './body'
 import imageGallery from './imageGallery'
 import buttonType from './buttonType'
+import youTube from './youTube'
 
-export const schemaTypes = [blogType, authorType, categoryType, body, imageGallery, buttonType]
+export const schemaTypes = [
+  blogType,
+  authorType,
+  categoryType,
+  body,
+  imageGallery,
+  buttonType,
+  youTube,
+]